--- a/test/test_conv.py
+++ b/test/test_conv.py
@@ -68,8 +68,8 @@ class TestConv(unittest.TestCase):
   def test_two_overlapping_binops_no_rerun_wino(self):
     Tensor.no_grad = True
     Tensor.wino = True
-    x = Tensor.randn(1,12,128,256)
-    w = Tensor.randn(32,12,3,3)
+    x = Tensor.randn(1,4,16,16)
+    w = Tensor.randn(6,4,3,3)
     out = x.conv2d(w, padding=(1,1))
     r1, r2 = out.relu(), out.elu()
     np.testing.assert_allclose(r1.numpy(), np.maximum(out.numpy(), 0))