--- a/tinygrad/codegen/optimizer.py
+++ b/tinygrad/codegen/optimizer.py
@@ -326,8 +326,9 @@ class OptimizedKernel(Kernel):
     # if there are small dims with lots of valid masks, upcast them (they might be from Tensor.stack)
     # this can be made much smarter
     to_upcast = []
+    # upcast leading axes first (hack-ish for winograd; we actually want to upcast masked axes with low stride first)
     for axis in range(self.first_reduce):
-      # todo: we need to be able to split axes that are masked, or refuse to merge them in simplify_merge_adjacent
+      # we might want to be able to split axes that are masked, or refuse to merge them in simplify_merge_adjacent
       if self.full_shape[axis] <= 36 and any(st.axis_needs_valid(axis) for st in self.sts) and prod(self.full_shape[self.shape_len-self.upcasted:]) * self.full_shape[axis] <= 7 * 7:
         if DEBUG >= 4: print(f"upcasting masked axis : {axis}")
         to_upcast.append(axis)