--- a/tinygrad/shape/shapetracker.py
+++ b/tinygrad/shape/shapetracker.py
@@ -196,6 +196,10 @@ class ShapeTracker:
   def needs_valid(self) -> bool:
     return any(v.mask is not None for v in self.views)
 
+  def axis_needs_valid(self, axis) -> bool:
+    idx, valid = self.expr_idxs()
+    return f'idx{axis}' in [v.expr for v in valid.vars()]
+
   # *** under this line are the movement ops ***
 
   def __unsafe_resize(self, arg: Tuple[Tuple[int, int], ...], mask=None):