New pr, old: #1037 . Tracking issue: #1618 . This is the minimal set of changes for fast winograd. Compilation time is slow with only this patch, that needs #1484 and some expr_idxs changes.

On rtx3090:

```
conv bs:  4 chans:128 -> 128 k:3              0.70 ms (86888.78 GFLOPS 126888.59 GB/s) in torch,    0.79 ms (77287.86 GFLOPS 112867.82 GB/s) in tinygrad,    1.12x slower   60806.07 MOPS 88798.54 MB
```

It can be 0.77ms with a slightly different hand coded optimization change, or it can be very good with KOPT.

We can merge once this part of the code is refined and the compile time is sufficiently low.