@tinyb0t 2023-08-26
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/codegen/optimizer.py             30               9
tinygrad/mlops.py                          4               0
tinygrad/shape/shapetracker.py             4               0
tinygrad/tensor.py                        32               4
------------------------------------------------------------
total                                     70              13
------------------------------------------------------------
lines added in the tinygrad folder: 57
```