@geohot 2023-08-30
I see, tests are failing because it's slow.

I should be able to finish up the UAst stuff by the end of the week, that'll be a much better starting point for making the linearizer fast.