@geohot 2023-09-01
Want to merge this with:
* winograd off by default
* enabled with WINO=1
* a few small tests to ensure correctness, okay if they take 30 seconds to compile a kernel

This way I can start playing with making the Linearizer fast in the tokenless branch