@chaosagent 2023-09-02
OK, two rudimentary tests are in. Maybe another test for no exploding kernel fusions might be nice though.