@geohot 2023-09-03
I understand the compile time is slow, but this is still slow in execution on M1:

With WINO=1 (something clearly broken in tracker):
`conv bs:  4 chans:128 -> 128 k:3              2.08 ms (29192.17 GFLOPS 42630.98 GB/s) in torch,    3.51 ms (17302.10 GFLOPS 25267.23 GB/s) in tinygrad,    1.69x slower   60806.07 MOPS 88798.54 MB`

Without:
`conv bs:  4 chans:128 -> 128 k:3              2.25 ms ( 8597.74 GFLOPS    30.59 GB/s) in torch,    4.22 ms ( 4583.83 GFLOPS    16.31 GB/s) in tinygrad,    1.88x slower   19327.35 MOPS    68.76 MB`