@geohot 2023-09-03
Hmm, I ran again with `DEBUG=4 WINO=1 MPS=1 BIG=1 python3 test/test_speed_v_torch.py TestBigSpeed.test_large_conv_3x3` and just this one: `conv bs:  4 chans:128 -> 128 k:3              2.97 ms (20441.30 GFLOPS 29851.59 GB/s) in torch,    2.46 ms (24686.55 GFLOPS 36051.17 GB/s) in tinygrad,    0.83x faster   60806.07 MOPS 88798.54 MB` that's a lot better.